Fixed more Inventory Stuff
-Leaving the bag and returning to the bag will reset the cursor to the top of the bag
-I cannot register items in the favorites pocket
-Using the Deep Sea Tooth or Magmarzier or any TM crashes the game.
-Using a Repel crashes the game
-Filling the TM pocket and trying to open the Bag crashes the game

+++ include/ui_inventory.h
@@ -18,9 +18,9 @@ struct StringList
 struct NewInventoryData
 {
     u8 numFavoriteItems[POCKETS_COUNT];
-    u8 itemIdx;
+    u16 itemIdx;
     u8 pocketNum;
-    u8 yFirstItem;
+    u16 yFirstItem;
     u16 registeredItem[NUM_REGISTER_ITEMS];
 };
